+++ src/docSnapshotCreator.ts
@@ -1,9 +1,9 @@
-import { Firelord } from './firelord'
+import { FirelordUtils } from './firelordUtils'
 import { FirelordFirestore } from './firelordFirestore'
 import { docCreator, DocCreator } from './docCreator'
 
 export const docSnapshotCreator: <
-	T extends Firelord.MetaType,
+	T extends FirelordUtils.MetaType,
 	M extends 'col' | 'colGroup' = 'col'
 >(
 	firestore: FirelordFirestore.Firestore,
@@ -13,8 +13,8 @@ export const docSnapshotCreator: <
 		? undefined
 		: never,
 	documentSnapshot: FirelordFirestore.DocumentSnapshot
-) => ReturnType<DocSnapshotCreator<T, M>> = <
-	T extends Firelord.MetaType,
+) => DocSnapshotCreator<T, M> = <
+	T extends FirelordUtils.MetaType,
 	M extends 'col' | 'colGroup' = 'col'
 >(
 	firestore: FirelordFirestore.Firestore,
@@ -25,7 +25,7 @@ export const docSnapshotCreator: <
 		: never,
 	documentSnapshot: FirelordFirestore.DocumentSnapshot
 ) => {
-	type Read = Firelord.InternalReadWriteConverter<T>['read']
+	type Read = FirelordUtils.InternalReadWriteConverter<T>['read']
 
 	return {
 		createTime: documentSnapshot.createTime,
@@ -53,23 +53,15 @@ export const docSnapshotCreator: <
 }
 
 export type DocSnapshotCreator<
-	T extends Firelord.MetaType,
+	T extends FirelordUtils.MetaType,
 	M extends 'col' | 'colGroup' = 'col'
-> = (
-	firestore: FirelordFirestore.Firestore,
-	colRef: M extends 'col'
-		? FirelordFirestore.CollectionReference
-		: M extends 'colGroup'
-		? undefined
-		: never,
-	documentSnapshot: FirelordFirestore.DocumentSnapshot
-) => {
+> = {
 	createTime?: FirelordFirestore.Timestamp
 	updateTime?: FirelordFirestore.Timestamp
 	exists: boolean
 	id: string
 	readTime: FirelordFirestore.Timestamp
-	ref: ReturnType<ReturnType<DocCreator<T, M>>>
+	ref: ReturnType<DocCreator<T, M>>
 	data: () => T['read'] | undefined
 	get: <F extends string & keyof T['write']>(fieldPath: F) => T['read'][F]
 	isEqual: (other: FirelordFirestore.DocumentSnapshot) => boolean