+++ src/dontRunThis.test.ts
@@ -4,8 +4,8 @@ import {
 	OmitKeys,
 	PartialNoImplicitUndefinedAndNoExtraMember,
 	ExcludePropertyKeys,
-	Firelord,
-} from './firelord'
+	FirelordUtils,
+} from './firelordUtils'
 import { firestore } from 'firebase-admin'
 import { flatten } from './utils'
 import { Wrapper } from './index_'
@@ -18,12 +18,12 @@ const {
 } = firelord(firestore)
 
 // use base type to generate read and write type
-type User = Firelord.ReadWriteCreator<
+type User = FirelordUtils.ReadWriteCreator<
 	{
 		name: string
 		age: number
 		birthday: Date
-		joinDate: Firelord.ServerTimestamp
+		joinDate: FirelordUtils.ServerTimestamp
 		beenTo: ('USA' | 'CANADA' | 'RUSSIA' | 'CHINA')[]
 	}, // base type
 	'Users', // collection path type
@@ -74,10 +74,10 @@ const userGroup = userCreator.colGroup('Users') // collection path type is "User
 const user = users.doc('1234567890') // document path is string
 
 // subCollection of User
-type Transaction = Firelord.ReadWriteCreator<
+type Transaction = FirelordUtils.ReadWriteCreator<
 	{
 		amount: number
-		date: Firelord.ServerTimestamp
+		date: FirelordUtils.ServerTimestamp
 		status: 'Fail' | 'Success'
 	}, // base type
 	'Transactions', // collection path type
@@ -385,7 +385,7 @@ users
 	.limit(1)
 	.where('beenTo', 'in', [['USA']])
 
-type a = Firelord.ReadWriteCreator<
+type a = FirelordUtils.ReadWriteCreator<
 	{
 		a:
 			| string
@@ -403,7 +403,7 @@ type b = a['write']
 type c = a['read']
 type f = a['compare']
 
-type a1 = Firelord.ReadWriteCreator<
+type a1 = FirelordUtils.ReadWriteCreator<
 	{
 		a: string | Date
 		b: { c: 1; d: 2 }
@@ -422,7 +422,7 @@ type x = Nested['read']
 type y = Nested['write']
 type z = Nested['compare']
 
-type Nested = Firelord.ReadWriteCreator<
+type Nested = FirelordUtils.ReadWriteCreator<
 	{
 		a: number
 		b: { c: string }
@@ -474,10 +474,10 @@ type NestedWrite = Nested['write']['d.e.g.h'] // {a: number | FirebaseFirestore.
 // compare type
 type NestedCompare = Nested['compare'] // {a: number, "b.c": string, "d.e.f": (FirebaseFirestore.Timestamp | Date)[], "d.e.g.h": FirebaseFirestore.FieldValue | { i: {j: firestore.Timestamp | Date}[] }[], createdAt: Date | firestore.Timestamp, updatedAt: Date | firestore.Timestamp}
 
-type Example = Firelord.ReadWriteCreator<
+type Example = FirelordUtils.ReadWriteCreator<
 	{
 		aaa: number | undefined
-		bbb: Firelord.ServerTimestamp
+		bbb: FirelordUtils.ServerTimestamp
 		ddd: string[]
 		eee: {
 			fff: {